# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SAMBINA",
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://sabina.dev:/api"
        }
    ],
    "paths": {
...
        "/forget-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Forget Password",
                "description": "Forget Password",
                "operationId": "forgetpassword",
                "requestBody": {
                    "description": "Forget Password",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "adminsambina@mail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "We have e-mailed your password reset link!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "mobilekey": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "token": {
                "type": "http",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "mobilekey": {
                "type": "apiKey",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    }
}